<description: restricted, no permission to view>
[CL 4287050 by Ben Marsh in Dev-Build branch]

+++ Engine/Source/Editor/BlueprintGraph/Classes/K2Node_InputTouchEvent.h
@@ -20,13 +20,13 @@ class UK2Node_InputTouchEvent : public UK2Node_Event
 	TEnumAsByte<EInputEvent> InputKeyEvent;
 
 	UPROPERTY()
-	uint8 bConsumeInput:1;
+	uint32 bConsumeInput:1;
 
 	UPROPERTY()
-	uint8 bExecuteWhenPaused:1;
+	uint32 bExecuteWhenPaused:1;
 
 	UPROPERTY()
-	uint8 bOverrideParentBinding:1;
+	uint32 bOverrideParentBinding:1;
 
 	//~ Begin UK2Node Interface
 	virtual UClass* GetDynamicBindingClass() const override;
